+++ pareg_core/src/check/mod.rs
@@ -1,0 +1,5 @@
+mod check_ref;
+mod in_range;
+mod in_range_i;
+
+pub use self::{check_ref::*, in_range::*, in_range_i::*};